Change the DEV/WINDEVICE output (the part that shows the mapping from dnet names to WinPcap names) to use the no-nonsense approach of simply calling DnetName2PcapName for every interface returned by getinterfaces. DnetName2PcapName calls intf_get_pcap_devname, which is same function called by eth_open on Windows, so now the output really reflects Nmap's view of the world.
This doesn't remove interface aliases, so if you have any aliases
enabled, you will see what look like duplicated lines. I have chosen to
leave them in for now to assist with debugging if any alias-related
problems come up in the future. They could be taken out easily if this
is not desired.

This brings what may be a startling change: the lo0 loopback interface
no longer maps to /Device/NPF_GenericDialupAdapter, but to no adapter at
all (signified by "<none>"). I believe the old behavior was wrong,
merely an artifact of the way the code happened to line up two lists of
interfaces. /Device/NFP_GenericDialupAdapter is for dialup and VPN
capture, and the loopback interface isn't supported by WinPcap (or
Windows for that matter) at all. See
http://www.winpcap.org/misc/faq.htm#Q-5
http://www.winpcap.org/misc/faq.htm#Q-13


git-svn-id: svn://svn.insecure.org/nmap@8135 e0a8ed71-7df4-0310-8962-fdc924857419