[NSE] This patch is related to the change to eliminate the reliance on GC for collecting socket locks [1].
If a thread does not close any sockets it creates, and then
returns (or errors), the thread and sockets will keep their "lock" until
garbage collected. This would be the same situation as before in this
particular case (reliance on GC).

To fix this, I have changed the socket unlock system to close all the sockets
of a thread not yielded and remove its "lock".

[1] http://seclists.org/nmap-dev/2009/q2/0624.html


git-svn-id: svn://svn.insecure.org/nmap@14002 e0a8ed71-7df4-0310-8962-fdc924857419