Remove extraports state inference from Ndiff. This was a feature that, if there was only one extraports state, would create a Port object for each scanned port that was not listed explictly. So for example, if a scan of 100 ports had 5 open ports and an extraports of "filtered", Ndiff would create 95 port records in the "filtered" state in addition to the 5 "open". If there was more than one extraports state this was not possible.
This turned out to be a slow operation, out of all proportion to its
utility. Here are times for a diff of random-1.xml and random-2.xml,
before and after:

real    0m11.912s
user    0m10.970s
sys     0m0.249s

real    0m0.773s
user    0m0.726s
sys     0m0.046s

In most cases extraports are not shown in the output, so this was wasted
effort. I know of only one place where it affected the output, when an
extraports in the A scan because a non-extraports in a different state
in the B scan. Then the previous state would be included in the <a>
port, where now it will just be listed without a state.


git-svn-id: svn://svn.insecure.org/nmap@14399 e0a8ed71-7df4-0310-8962-fdc924857419